requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
orjson==3.10.7
pydantic==2.9.2
pydantic-settings==2.6.0
SQLAlchemy==2.0.36
alembic==1.13.2
psycopg2-binary==2.9.9
httpx==0.27.2
redis==5.0.8
rq==1.16.2
boto3==1.35.24
python-dotenv==1.0.1
structlog==24.4.0
sentry-sdk==2.17.0
python-multipart==0.0.12
prometheus-client==0.21.0
requests==2.32.3
fal-client>=0.4.0

# OAuth
Authlib==1.3.1

# Starlette session signing
itsdangerous==2.2.0
